upload_service/requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-pubsub
# google-cloud-firestore
# google-cloud-storage
# google-cloud-pubsub
# google-cloud-logging
# PyJWT